Broke out Buzzer, LED, and Output code into classes in there own files. Currently the new implementation of Output doesn't work so it's commented out in the main file.

+++ math.h
@@ -15,6 +15,6 @@ int CelsiusToFahrenheit(float temp) {
   return (temp * (9/5)) + 32;
 }
 
-int MetersToFeet(float meters) {
+short MetersToFeet(float meters) {
   return meters * 3.28084;
 }
